refactor random loader choice into CustomLoadersProvider component
